Run girder in --dev mode with ocdeploy
Utilize the new --dev option for girder so that it can be ran in
development mode. This is useful because in development mode, girder
will automatically restart if changes to the plugins are made.

If the GIRDER_EXTRA_ARGS environment variable is not set or empty, then
girder is ran in production mode.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>