Added script to upload molecules from qcarchive
The script contains a few variables at the beginning that
define the qcarchive collection from which to obtain the
molecules. The default is the "JGI Metabolite Set 1"
OptimizationDataset.

The qcarchive molecule id is stored in the "provenance", so
that a user can refer back to the molecule in the qcarchive
if they want more information.

For some bigger molecules, it takes a very long time (>1 minute) to
upload, and the mongochemclient will be unable to communicate
with the girder server during this time. We should probably figure
out why some of them take so long.

Depends on [this PR](https://github.com/OpenChemistry/mongochemserver/pull/142).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>