Added basic Open Babel flask container
It can currently only convert formats, but more endpoints and
functions can be added in the future.

Documentation on the convert function can be seen in the main.py
file.

An example of using it to convert smiles to xyz and generate 3D
coordinates is as follows:
```
> curl -X POST 'http://localhost:5000/convert/xyz' -H "Content-Type: application/json" -d '{"input_format": "smiles", "data": "CCO", "gen3d": true}'

9

C          1.07579        0.04200       -0.05942
C          2.59038        0.03826       -0.05539
O          3.07166       -1.22565        0.37800
H          0.68946        1.00611       -0.40166
H          0.68523       -0.15819        0.94373
H          0.69041       -0.74634       -0.71434
H          2.98292        0.81128        0.61217
H          2.97618        0.22333       -1.06208
H          2.73371       -1.37537        1.27738
```

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>