Use Open Babel flask docker image
In mongochemserver, as of [this PR](https://github.com/OpenChemistry/mongochemserver/pull/144),
Open Babel is no longer used through direct API calls. Instead,
it is called as an Open Babel flask docker image.

This PR gets mongochemdeploy to add Open Babel to docker-compose,
and the config is set up so that it can be called by mongochemserver.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>