Fix the torchani docker build and lock in versions
The docker build was broken because "torchvision_nightly" no longer
exists on pypi.

Rather than downloading all of the latest libraries for torchani, lock
in several of the libraries' versions so we can build a consistent
image. This also enables us to keep the torchani version in the
description.json file up-to-date.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>