Add cp2k docker image for energy and optimizations
This adds a cp2k docker image that can perform energy and
optimization calculations. It currently does not produce
any orbitals, but it can calculate the energy of a system and
perform a geometry optimization.

Since cp2k is primarily designed for periodic systems, the
calculations are performed by placing the molecule in the
center of a unit cell with at least 10 angstroms of distance
between atoms in neighboring unit cells. This is done so that
the atoms in neighboring unit cells will (hopefully) not interact.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>